Fix Mercurial unable to authenticate when pulling
In some configuration, Diffusion is unable to update because Mercurial fails to provide the necessary credentials during the pull command.

I understand that HTTP credentials are passed in the modified remote URL (using username:password@host scheme), but when calling hg pull, the remote URL is not given again. Mercurial only stores the username in the remote URL, so to be able to authenticate, we need to give it the full remote URL again.